+++ lib/astro/lunar.ex
@@ -1,2 +1,4 @@
 defmodule Astro.Lunar do
+  @moduledoc false
+
 end